requirements.txt
anthropic>=0.40.0
jupyter
notebook
ipykernel
python-dotenv
httpx
pydantic>=2.0
chromadb
sentence-transformers
pandas
numpy
matplotlib
seaborn
rich
tiktoken
mcp
